backend/requirements.txt
anyio==4.9.0
asgiref==3.8.1
channels==4.2.2
click==8.2.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-extra-fields==3.7.0
filetype==1.2.0
h11==0.16.0
httptools==0.6.4
idna==3.10
pillow==11.2.1
PyJWT==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
qrcode==8.2
rest-framework-simplejwt==0.0.2
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
